requirements.txt
flask>=2.0.0
openslide-python>=1.2.0
Pillow>=9.0.0
